+++ render.go
@@ -19,6 +19,9 @@ type Render struct {
 	row                uint16
 	col                uint16
 
+	reverseSearchEnabled bool
+	searchPrefixIsSet    bool
+
 	previousCursor int
 
 	// colors,
@@ -39,6 +42,13 @@ type Render struct {
 	scrollbarThumbColor          Color
 	scrollbarBGColor             Color
 }
+
+const (
+	matchSearchPrefix = "bck-i-search"
+	failSearchPrefix  = "failing bck-i-search"
+	maxBackwardColumn = 256
+	magic             = 5
+)
 
 // Setup to initialize console output.
 func (r *Render) Setup() {
@@ -175,7 +185,7 @@ func (r *Render) ClearScreen() {
 }
 
 // Render renders to the console.
-func (r *Render) Render(buffer *Buffer, previousText string, completion *CompletionManager) {
+func (r *Render) Render(buffer *Buffer, previousText string, completion *CompletionManager, history *History) {
 	defer debug.Un(debug.Trace("Render"))
 	// In situations where a pseudo tty is allocated (e.g. within a docker container),
 	// window size via TIOCGWINSZ is not immediately available and will result in 0,0 dimensions.
@@ -212,13 +222,18 @@ func (r *Render) Render(buffer *Buffer, previousText string, completion *Complet
 
 	r.out.EraseLine()
 	r.out.EraseDown()
-	r.renderPrefix()
 
-	if buffer.NewLineCount() > 0 {
-		r.renderMultiline(buffer)
+	if r.reverseSearchEnabled {
+		r.renderReverseSearch(buffer, history)
 	} else {
-		r.out.WriteStr(line)
-		defer r.out.ShowCursor()
+		r.renderPrefix()
+
+		if buffer.NewLineCount() > 0 {
+			r.renderMultiline(buffer, history)
+		} else {
+			r.out.WriteStr(line)
+			defer r.out.ShowCursor()
+		}
 	}
 
 	r.lineWrap(cursor)
@@ -245,7 +260,7 @@ func (r *Render) Render(buffer *Buffer, previousText string, completion *Complet
 	r.previousCursor = cursor
 }
 
-func (r *Render) renderMultiline(buffer *Buffer) {
+func (r *Render) renderMultiline(buffer *Buffer, history *History) {
 	before := buffer.Document().TextBeforeCursor()
 	cursor := ""
 	after := ""
@@ -263,12 +278,38 @@ func (r *Render) renderMultiline(buffer *Buffer) {
 
 	r.out.SetColor(r.inputTextColor, r.inputBGColor, false)
 	r.out.WriteStr(before)
+	history.Add(before)
 
 	r.out.SetDisplayAttributes(r.inputTextColor, r.inputBGColor, DisplayReverse)
 	r.out.WriteStr(cursor)
 
 	r.out.SetColor(r.inputTextColor, r.inputBGColor, false)
 	r.out.WriteStr(after)
+}
+
+func (r *Render) renderReverseSearch(buffer *Buffer, history *History) {
+	if !r.searchPrefixIsSet {
+		r.renderPrefix()
+		r.out.WriteStr(fmt.Sprintf("\n%s: %s", matchSearchPrefix, buffer.Text()))
+		r.searchPrefixIsSet = true
+	} else {
+		upLinesCnt := strings.Count(history.lastReverseFinded, "\n") + 1
+		searchPrefix := matchSearchPrefix
+		if !history.reverseFindInHistory(buffer.Text()) {
+			searchPrefix = failSearchPrefix
+		}
+
+		r.out.CursorBackward(magic)
+		for i := 0; i < upLinesCnt; i++ {
+			r.out.CursorUp(1)
+			r.out.EraseLine()
+		}
+
+		r.renderPrefix()
+		r.out.WriteStr(fmt.Sprintf("%s\n%s: %s", history.lastReverseFinded, searchPrefix, buffer.Text()))
+	}
+
+	r.out.ShowCursor()
 }
 
 // BreakLine to break line.