+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Change Log
 
-## v0.3.0 (2018/??/??)
+## Unreleased
 
-next release.
+### Added
+
+* Reverse search
 
 ## v0.2.3 (2018/10/25)